+++ lib/number.dart
@@ -1,5 +1,6 @@
 /// Support for fractional and complex arithmetic.
 export 'src/number/complex.dart' show Complex;
 export 'src/number/fraction.dart' show Fraction;
-export 'src/number/mixins/close_to.dart';
+export 'src/number/mixins/close_to.dart' show CloseTo;
+export 'src/number/mixins/close_to_extension.dart' show CloseToNumExtension;
 export 'src/number/quaternion.dart' show Quaternion;